refactor utils into utils
